feat: introduce an init module which takes care of setting up app config and logging
