+++ terraform_drift_detection/detect.py
@@ -2,11 +2,11 @@
 
 import logging
 
+from terraform_drift_detection.init import init
 from terraform_drift_detection.terraform import check_repos
-from terraform_drift_detection.util import Drift, setup_logging
+from terraform_drift_detection.util import Drift
 
-setup_logging()
-results = check_repos()
-if Drift.DRIFT in results:
+init()
+if Drift.DRIFT in check_repos():
   logging.error('Drift detected.')
   exit(1)
